+++ src/scripts/background/interface/alert/manageModals.js
