+++ src/scripts/background/interface/alert/managePopups.js
