+++ src/scripts/background/interface/alert/setUninstalledRedirect.js
