+++ src/scripts/background/interface/contextMenu/buildContextMenu.js
