+++ src/scripts/background/interface/icon/clickCount.js
@@ -1,6 +1,6 @@
 //TODO: object Clicker with array of numbers : each time click on icon, use the next number before displaying popup
 
-import {getStorageData} from "../../shared/utils/chromeStorage";
+import {getStorageData} from "../../../shared/utils/chromeStorage";
 
 // Array of click counts for each modal
 export const clickCounts = [14, 25, 40, 40] //[20, 40, 60, 80, 100];
