+++ src/scripts/background/interface/icon/iconAction.js
